+++ spring-dataflow-ms-sink/create-dataflow-sink-server-docker-image.sh
@@ -15,8 +15,19 @@ if [ -e dataflow-ms-flow-centric-sink-poc.jar ]; then
 	rm -f dataflow-ms-flow-centric-sink-poc.jar
 fi
 
-cp $JAVA_PROJECT_FOLDER/dataflow-ms-flow-centric-sink-poc/target/dataflow-ms-flow-centric-sink-poc-*.jar dataflow-ms-flow-centric-sink-poc.jar 
+
+
+cp $JAVA_PROJECT_FOLDER/dataflow-ms-flow-centric-sink-poc/target/dataflow-ms-flow-centric-sink-poc-*.jar ./dataflow-ms-flow-centric-sink-poc.jar 
+
+if [ -e ./dataflow-ms-flow-centric-sink-poc.jar   ]; then
 
 docker --debug image build --rm . -t $DOCKER_IMAGE_USER/spring-cloud-dataflow-sink-server:$SERVER_RELEASE
 
 rm dataflow-ms-flow-centric-sink-poc.jar
+
+else
+	echo "Unable to copy file $JAVA_PROJECT_FOLDER/dataflow-ms-flow-centric-sink-poc/target/dataflow-ms-flow-centric-sink-poc-*.jar"
+	echo "From  folder $JAVA_PROJECT_FOLDER/dataflow-ms-flow-centric-sink-poc/target, listed as:"
+	echo "Files: $(ls $JAVA_PROJECT_FOLDER/dataflow-ms-flow-centric-sink-poc/target)"
+	exit 1
+fi
